Renamed App, package and namespace
