+++ app/src/main/java/io/github/ferreol/seteditplus/Native.java
@@ -1,4 +1,4 @@
-package io.github.muntashirakon.setedit;
+package io.github.ferreol.seteditplus;
 
 import android.os.Build;
 import java.util.Collections;
