+++ app/src/main/java/io/github/ferreol/seteditplus/PropertyCallback.java
@@ -1,4 +1,4 @@
-package io.github.muntashirakon.setedit;
+package io.github.ferreol.seteditplus;
 
 public interface PropertyCallback {
     void handleProperty(String key, String value);
