+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name" translatable="false">Settings Database Editor</string>
+    <string name="app_name" translatable="false">Settings Database Editor+</string>
     <string name="save">Save</string>
     <string name="error_no_support"><b>SetEdit</b> does not currently support editing this table. To enable editing, run the following command in a root or ADB shell:</string>
     <string name="error_unexpected"><b>Unexpected failure:</b></string>
@@ -27,5 +27,5 @@
     <string name="theme_battery">Follow battery saver</string>
     <string name="theme">Theme</string>
     <string name="shortcut">Shortcut</string>
-    <string name="shortcut_not_supported">Shortcut not suported!</string>
+    <string name="shortcut_not_supported">Shortcut not supported!</string>
 </resources>
